+++ src/components/MGRSInput.js
@@ -31,6 +31,8 @@
 // Top Level Modules
 //----------------------------------------------------------------//
 import React from 'react'
+import Dms from 'geodesy/dms'
+import Mgrs, { LatLon } from 'geodesy/mgrs'
 
 //----------------------------------------------------------------//
 // Material-UI Core Components
@@ -88,8 +90,92 @@ const useStyles = makeStyles(theme => ({
 //----------------------------------------------------------------//
 // MGRS Input Component
 //----------------------------------------------------------------//
-export default () => {
+export default ({ map, setClickedLatLng }) => {
   const classes = useStyles()
+
+  let inputRef = React.useRef('')
+  const [inputValue, setInputValue] = React.useState('')
+
+  /**
+   * 
+   */
+  const handleSubmit = () => {
+    // As long as the input is not blank
+    if (inputValue !== '') {
+      inputRef.blur()
+
+      // RemoveSpaces
+      let target = inputValue
+      let validInput = true
+
+      let letterCount = 0
+      for (let i = 0; i < target.length; i++) {
+        if (target.charAt(i) !== ' ' && target.charAt(i) !== '.' && !Number.isInteger(parseInt(target.charAt(i)))) {
+          letterCount++
+        }
+      }
+
+      // Determine if use input MGRS or Lat Lon and then fly to the location
+      if (letterCount < 2) {
+        console.error('Error: Invalid Lat/Lon, N/S and E/W are required.')
+        validInput = false
+      } else if (letterCount < 3) {
+
+        // Find the end of the latitude
+        let index = target.indexOf('N')
+        if (index === -1) {
+          index = target.indexOf('S')
+        }
+
+        // Try to parse lat and lng
+        if (index !== -1) {
+          target = LatLon.parse(target.substr(0, index + 1).trim(), target.substr(index + 1).trim())
+        } else {
+          validInput = false
+        }        
+      } else {
+        // Remove spaces
+        target = target.replace(/ /g, '')
+
+        // Add leading zero
+        if (!Number.isInteger(parseInt(target.charAt(1)))) {
+          target = '0' + target
+        }
+
+        // Split out the sections
+        const grid = target.substr(0, 5)
+        const digits = target.substr(5)
+        let easting = digits.substr(0, digits.length / 2)
+        let northing = digits.substr(digits.length / 2)
+
+        // Pad easting/northing
+        while (easting.length < 5) {
+          easting = easting + '0'
+        }
+
+        while (northing.length < 5) {
+          northing = northing + '0'
+        }
+
+        // Recombine gridzone, easting, and northing
+        target = grid + easting + northing
+        try {
+          target = Mgrs.parse(target)
+          target = target.toUtm().toLatLon()
+        } catch (error) {
+          console.error(error)
+          validInput = false
+        }
+      }
+
+      if (validInput) {
+        map.flyTo([target.lat, target.lon], 10)
+        setClickedLatLng({ lat: target.lat, lng: target.lon })
+      } else {
+        console.error(`Error: Invalid coordinates {${target}}`)
+      }
+    }
+  }
 
   return (
     <div className={classes.search}>
@@ -101,13 +187,11 @@ export default () => {
           input: classes.inputInput,
           root: classes.inputRoot,
         }}
-        onChange={event => console.log(event.target.value)}
-        onKeyDown={event => {
-          if (event.key === 'Enter') {
-            console.log('Enter Pressed - how to blur this?')
-          }
-        }}
+        inputRef={input => inputRef = input}
+        onChange={input => setInputValue(input.target.value.toUpperCase())}
+        onKeyDown={event => (event.key === 'Enter') ? handleSubmit() : undefined}
         placeholder='MGRS or Lat Long'
+        value={inputValue}
       />
     </div>
   )